pyproject.toml
[tool.poetry]
name = "NiceToMeetYou"
version = "0.1.0"
description = ""
authors = ["YuZhan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0"
celery = "5.3.6"
redis = "5.0.1"
flower = "2.0.1"
psycopg2-binary = "2.9.9"
watchfiles = "0.21.0"
requests = "2.31.0"
daphne = "4.0.0"
channels = "4.0.0"
channels-redis = "4.1.0"
django-celery-beat = "2.6.0"
factory-boy = "3.3.0"
coverage = "7.4.0"
pytest = "7.4.4"
pytest-django = "4.7.0"
pytest-factoryboy = "2.6.0"
pytest-cov = "4.1.0"
pillow = "10.1.0"
gunicorn = "21.2.0"
uvicorn = {version = "0.25.0", extras = ["standard"]}
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
